requirements.txt
streamlit
pandas
matplotlib
seaborn
numpy
scikit-learn
sentence-transformers
transformers
torch
arxiv
requests
python-dotenv
groq
PyMuPDF
